+++ pages/account/manage/[shopID]/orders.tsx
@@ -2,7 +2,7 @@
 import PageHeader from "@/components/PageHeader";
 import ManageShopTabs from "@/components/account/manage/ManageShopTabs";
 import OrderListItem from "@/components/account/manage/OrderListItem";
-import OrderStatusSelect from "@/components/account/manage/OrderStatusSelector";
+import OrderStatusSelect from "@/components/account/manage/OrderStatusSelect";
 import SkeletonOrderListItem from "@/components/account/manage/SkeletonOrderListItem";
 import cn from "@/utils/helpers/cn";
 import createJimmy from "@/utils/helpers/createJimmy";
